working towards running any process
Hit a road block with __rte() giving hard faults. This is something
wrong with the process setup, I'm guessing. Or the stack allocation, but
I really feel as though I've done a good job fixing that. Lots of
changes to the bounds of the heap. Added a macro for the STACK_SIZE
since it's hard-coded in usr_proc.c i believe.